Added the ClipViewPort function and added some documentation notes to the GlUtilities class.
Added:
A ClipViewPort method that helps create a clipping port inside of the Unity editor to cull out any graphics library, GPU command culling, or handle rendering that would be out of the defined clip rect that is passed into this function.

Example of this function being used is inside of the SF Sprite Editor to clip and graphics rendering when users are dragging out a new sprite rect and accidentally move the mouse out of the sprite editor window. It also helps clip graphic calls for instance rendering textures, sprites, and shader drawn virtual textures. This is used in the SF Sprite Editor to also keep the texture from rendering out of the window when the sprite sheet is panned partially off screen.

Due note in most cases for the sprite sheet part this method might not be needed if the sprite sheet is being drawn directly into a UI Visual Element instead of IMGUI or an IMGUI container. There is a USS style property called   --unityOverflowClipBox that can have it's value set to Content Box. Combine the -unityOverflowClipBox property with the property overflow having it's value set to hidden, you can do have the UI Render do the ClipViewPort culling for you automatically without doing it in C# code.