Made a RectExtensions class for extension methods for the struct type Rect. Added two methods for Flipping negative values of a Rect and getting it's area.
Added:
1. A new class called RectExtensions to help with Rect extension methods to ease the manipulation of values in Rects.
2. Made a method for making sure a Rect doesn't have any negative values for x,y, width, and height. Example of one case this can happen is documented below in the commit message.
3. Added a method to get the area of a Rect. Even though there is a size value for Rect's there is a chance Rects can have a negative width or height. This method first calculates the Abosolute values to make sure width and height are not negative than does the size calculation on the proper width and height values.

For a see able example when width and height can be negative you can look at the SF Sprite Editor inside the SF 2D Tools package.
When users are drawing a Rect to create a new Sprite in the SFSpriteEditor if they start with the top right of the rect and draw the sprite rect going from top right to bottom left the value of the Rect width is technically a negative width and height since the Rect is based on the starting mouse position where the sprite rect was starting to be drawn going toward the final mouse position when the user releases the mouse.