Added a WIP HandlesUtility class for testing out custom editor Handle functionality.
Important Note this file is a early WIP file. Only committing because the DrawCircle2D function that is already typed up is needed in a couple other Sf packages for things like character vision cone visual debugging and some soon to be released shadow masking functionality that can use a Handle for controlling the side of the shadow mask size in editor.

Added:
1. Handles Utility class to make certain Editor Handle features easier to work with. This includes custom handle types and handle cap render shapes like 2D arcs.
2. Added the DrawCircle2D function for making a 2D circle handle renderer. This was made because the built in Unity Circle Handles is based in 3D space and there is no built in ones for 2D space. The 2D space one is useful when doing adding Editor functionality for tools that are in 2D space like screen space coordinates in Editor Windows and inside of the Scene view for 2D based games.

Below was only pushed to let the first method be able to be used. Don't try using them yet. They are not yet implemented.
3. DrawArc2D will be similar to the DrawCircle2D, but it allows for only making a part of a circle. This can be useful for more fine tuned control for things like vision cones and direction checking without worrying about checking for something in certain directions.
4. GetCircleSectorPoints will be implemented in the future to get a Vector2 array, Vector2 list, or a Vector2 span for best performance.
The Vector2 collections will contain the cut out sectors in a circle. This can be useful for making the DrawArc2D function and other special shapes that are cut out of other normal shapes.

Important note. Sector and segments of a circle are two different things.

A circle is formed by the set of points that are at a constant or at a fixed distance (radius) from a fixed point (center) in the plane.
The area enclosed by two radii and the corresponding arc in a circle is called a sector.
Theta angle can be seen at the angle difference from the start of one radii of the sector to it's end point of a sector.
Example a semi circle is a Sector with a theta angle of 180%

Area of sector when using degrees for Theta = ( Theta Angle in Degrees / 360 ) * pi * radius squared
Area of sector when using Radians for Theta = ( 1 / 2 ) * radius squared * Theta Angle in Radians

Segment of a circle = The region that is bounded by an arc and a chord of the circle.
Area of segment = area of sector - area of triangle