Added an entire system for procedurally generating different types of meshs using the job system and burst compiler.
Credit to CatLikeCoding for teaching people how to do this.

Created a series of Mesh Jobs that implement an IMeshGenerator to procedurally generate different shapes.

IMeshGenerator defines the base implementation for executing a C# job that generates shapes procedurally.
It includes fields for Vertex and Index count for the generated mesh.

IMeshStream is used to initialize the C# Job for procedural data creation used in meshes like defining the Triangles, SetVertex positions and calculating the bounds.

MeshJob is used to schedule jobs defined in the IMeshGenerator being implemented by the chosen ProceduralMeshs and to also pass in the IMeshStream.