Read commit description. Added the start of the Shape Factory class, the SF MeshUtilities, and the SFRendererUtilities.
Note these static classes will be split up in the future into their own files.

The Shape Factory was originally attended for Mesh and Gizmos shapes, but with the new low level 2D physics, they will be heavily useful as I add more to them for multithreaded set up for custom collider shapes.

The SFMeshUtilities will be used later for creating Meshes for UI Toolkit, Shaders, and low level 2D physics.
This is going to be useful for creating 2D meshes for interactive enviormental effects:

1. Creating ice/earth pillars from the ground shapes to match their collision shape.
2. Freezing water to be able to step on the water.
3. Lot more honestly.

SFRendererUtilities is used for projection matrix helpers, ScriptableRenderContext used when rendering meshes with a list of cameras, and some useful editor only materials shotcuts.